Updates on health and rank module
These updates basically set the groundwork for the health system for my game and the ranking system. I basically decided to work on this while I figure out how to jump. Why is it so hard to jump. I really hope Unity and Unreal is not like this. Oh God.