Jumping still doesn't really work
For some reason jumping doesn't really work. If one just slightly presses the jump button. The player sprite goes up a little and goes down as intended but if on was to press the jump button again they will be able to go upwards again. which is not supposed to happen.